services/site/requirements.txt
aiohttp~=3.8.6
alembic~=1.12.1
asyncpg~=0.29.0
fastapi~=0.104.1
jinja2~=3.1.2
isort~=5.12.0
nats-py~=2.6.0
pydantic-settings~=2.1.0
pydantic[email]~=2.5.2
pydantic~=2.5.0
PyJWT~=2.8.0
python-multipart~=0.0.6
requests~=2.31.0
ruff~=0.1.6
SQLAlchemy~=2.0.23
uvicorn~=0.24.0.post1
greenlet~=3.0.2
itsdangerous~=2.1.2